Move files around and prepare notebooks and readme
