--- .dockerignore
@@ -1,12 +1,0 @@
-node_modules/
-lib/
-*ipynb_checkpoints
-*.lock
-yarn-error.log
-.cache
-.ipython
-.jupyter
-.local
-.node-gyp
-.npm
-__pycache__
