--- binder/environment.yml
@@ -1,14 +1,0 @@
-name: jupyterlab-omnisci
-channels:
-  - conda-forge
-dependencies:
-  - python
-  - jupyterlab
-  - altair
-  - pyyaml
-  - mypy
-  - pip:
-    - vdom
-    - vega_datasets
-    - pymapd
-    - git+https://github.com/ibis-project/ibis.git
