--- docker-compose.yml
@@ -1,16 +1,0 @@
-version: "3"
-services:
-  jupyter:
-    build: .
-    environment:
-      JUPYTER_ENABLE_LAB: 'true'
-      YARN_CACHE_FOLDER: /yarn-cache
-    command: start-notebook.sh --watch
-    ports:
-      - "8888:8888"
-    volumes:
-      - ".:/home/jovyan"
-      - "./docker-start.sh:/usr/local/bin/start-notebook.d/docker-start.sh"
-      - "yarn-cache:/yarn-cache"
-volumes:
-  yarn-cache:
