+++ jupyterlab_omnisci/__init__.py
@@ -1,0 +1,2 @@
+from .altair import *
+from .magics import *
